fix: you can now not delete journal entires you didn't create

+++ src/routes/api/journalEntry/+server.ts
@@ -12,16 +12,16 @@ export async function GET() {
     return new Response(JSON.stringify(entry));
 }
 
-export const POST = (async ({ locals}) => {
+export const POST = (async ({ locals }) => {
     try {
         const user = (await locals.validateUser()).user;
-        if(user) {
+        if (user) {
             const today = new Date();
             const dd = String(today.getDate()).padStart(2, '0');
             const mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
             const yyyy = today.getFullYear();
             const todayString = mm + '/' + dd + '/' + yyyy;
-        
+
             const newEntry = await prisma.journalEntry.create({
                 data: {
                     title: todayString,
@@ -29,15 +29,15 @@ export const POST = (async ({ locals}) => {
                     user_id: user.userId,
                 }
             });
-        
+
             const id = newEntry.id;
-        
+
             console.log(id);
-        
+
             return new Response(JSON.stringify({ id: id }), { status: 201 })
         }
         else return new Response(JSON.stringify({ message: "Unauthorized" }), { status: 401 })
-        
+
     } catch {
         return new Response(JSON.stringify({ message: "failure" }), { status: 400 })
     }
@@ -78,10 +78,24 @@ export const PATCH = (async ({ request }) => {
     return new Response(JSON.stringify({ message: "success" }), { status: 201 })
 }) satisfies RequestHandler;
 
-export const DELETE = (async ({ request }) => {
-    const body = await request.json()
+export const DELETE = (async ({ request, locals }) => {
+    const user = (await locals.validateUser()).user;
+    if (!user) {
+        return new Response(JSON.stringify({ message: "Unauthorized" }), { status: 401 });
+    }
 
+    const body = await request.json()
     const id = parseInt(body.id);
+
+    const entry = await prisma.journalEntry.findUnique({
+        where: {
+            id: id
+        }
+    });
+
+    if (!entry || entry.user_id !== user.userId) {
+        return new Response(JSON.stringify({ message: "Unauthorized" }), { status: 401 });
+    }
 
     const deleteEntry = await prisma.journalEntry.delete({
         where: {
